+++ tests/unit/domestic/test_models.py
@@ -34,6 +34,7 @@ from domestic.models import (
     industry_accordions_validation,
     main_statistics_validation,
 )
+from international_online_offer.models import IOOIndexPage
 from tests.helpers import SetUpLocaleMixin, create_response
 from tests.unit.core.factories import (
     CountryFactory,
@@ -1972,6 +1973,7 @@ class StructuralPageTests(WagtailPageTests):
             StructuralPage,
             {
                 ArticlePage,
+                IOOIndexPage,
             },
         )